src/ai-services/requirements.txt
flask>=2.0.0
openllm>=0.2.0
transformers
torch
accelerate
gunicorn
python-dotenv